pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PaperFarm"
version = "0.2.0b1"
description = "Let AI agents run experiments in any repo while you sleep."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PaperFarm Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
keywords = ["ai", "research", "experiments", "automation", "agent"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "textual>=0.85.0",
    "filelock>=3.12.0",
    "plotext>=5.3.0",
    "textual-plotext>=1.0.0",
]

[project.optional-dependencies]
tokens = ["tiktoken>=0.7.0"]
serve = ["textual-serve>=1.1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[project.scripts]
PaperFarm = "paperfarm.cli:app"
paperfarm = "paperfarm.cli:app"
open-researcher = "paperfarm.cli:app"

[project.urls]
Homepage = "https://github.com/shatianming5/PaperFarm"
Repository = "https://github.com/shatianming5/PaperFarm"
Issues = "https://github.com/shatianming5/PaperFarm/issues"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/paperfarm", "src/open_researcher"]
